otree-core-master/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgi-redis==0.11.0
asgiref==0.13.0
autobahn==0.14.0
Babel==2.3.4
boto==2.40.0
channels==0.13.1
colorama==0.3.7
contextlib2==0.5.3
coverage==4.0.3
daphne==0.12.0
dj-database-url==0.4.1
Django==1.8.8
django-countries==3.4.1
django-easymoney==0.7.1
django-extensions==1.6.7
django-floppyforms==1.6.2
django-idmap==0.4.1
django-inspect-model==0.7
django-vanilla-views==1.0.4
djangorestframework==3.3.3
funcsigs==1.0.2
honcho==0.7.1
huey==1.1.2
IPy==0.83
mock==2.0.0
msgpack-python==0.4.7
ordered-set==2.0.1
otree-save-the-change==1.1.3
pbr==1.9.1
pytest==2.9.2
pytest-django==2.9.1
pytz==2016.4
raven==5.17.0
redis==2.10.5
requests==2.10.0
schema==0.5.0
six==1.10.0
Twisted==16.1.1
txaio==2.5.1
unicodecsv==0.14.1
wheel==0.29.0
whitenoise==3.0
ws4py==0.3.5
xmltodict==0.10.1
zope.interface==4.1.3

# Stress test
selenium==2.53.2
psutil==4.2.0
tqdm==4.7.4

